+++ scripts/bin/react/_create.sh
@@ -65,7 +65,7 @@ function reactCreate() {
         $(voidSubstr '{{FiddleName}}' ${fiddle} "README.md";) || exit 5;
         $(voidSubstr '{{BornOnDate}}' ${bornOnDate} "README.md";) || exit 5;
         #createTypingsRcFile || exit 6;
-        npmShrinkWrap || exit $?;
+        #npmShrinkWrap || exit $?;
 
     )
     # catch